Updated openstack/openstack Project: openstack-infra/zuul  773651ad7bf0fc6adba2357173ffb657d874478a
Only depend-on open changes

Instead of checking if changes found from a Depends-On line are merged,
check if they are open. This ignores abandoned changes when building the
list of dependent changes.

In the event that change I0000 depends-on change-id I1234 and change-id
I1234 matches multiple changes, one of which is abandoned, then change
I0000 will never enter the gate as it waits indefinitely for the abandoned
I1234 to become eligible for the gate.

Change-Id: Ia6ae7b1da13fa2bcf6fc08fac49563e7ac960a84