Updated openstack/openstack Project: openstack-infra/shade  c47ec154578e505eba9c81a8c90ef82703c697af
Consider 'in-use' a non-pending volume for caching

We don't cache volume list if one of the volumes is in pending state, as
otherwise we'd miss the cloud-side cache invalidation. However, active
volumes that are attached are in state "in-use" - so basically if you
had one active volume, the cache would never cache.

Change-Id: I03cbc1b814e4a5829936a22751ee81d52b83fb2e