Updated openstack/openstack Project: openstack-infra/release-tools  5576ef168d6dda8916e63123251413e8330c6d57
make manual releases easier with option to generate announcement

Add -a option to release_from_yaml.sh and release.sh, causing the latter
to trigger announce.sh to generate the release announcement. This saves
a couple of extra manual steps, since the announce script can run
against the already cloned project repository with the new tag, instead
of having to do that step separately. The option is only meant to be
used when the scripts are run interactively.

Change-Id: I3449ca119d42c0317de2aa0fc5441091acfe49e1