Updated openstack/openstack Project: openstack/oslo.versionedobjects  371a445957afbd06bf8ae62042b3ef45065a3b6e
Add a callable to gather extra data for fps

ObjectVersionChecker gathers a group of relevant data with which it then
generates a fingerprint. This change adds a callable to get_hashes() to
provide the object class, which the callable can then use to gather more
data upon which it wants the fingerprint to be generated. This allows
end users to version on more than just fields and methods.

Change-Id: Ibbaaf926d50693ed0ddd96132542830dba75a285