Updated openstack/openstack Project: openstack/nova  46f829f3e31a154ce133cad5b3b33004184af557
neutron: only list ports if there is a quota limit when validating

The list_ports call can take awhile if the project has a lot of ports.
If it turns out that there is unlimited quota, then we don't even need
to list the ports, so move that after the show_quota call.

Partial-Bug: #1521788

Change-Id: I4d128f182283ffa4479934f640a67d9c536824b5