Updated openstack/openstack Project: openstack/nova  89464800103423d032b770d1ce6d1820c102d1a2
neutron: only get port id when listing ports in validate_networks

The validate_networks method is called from the compute API before
casting off to build instances. If the project has a lot of ports, the
list call to neutron can be slow, so we should optimize this method as
much as possible. One way to do that is rather than get the port details
back, we just get the id since all we care about is the number of ports
assigned to the project.

Closes-Bug: #1521788

Change-Id: Icddba161112526dde23184ee80ad5cb60fe66cc6