Updated openstack/openstack Project: openstack-infra/system-config  a9e8854c1345af6858c7a7e92c2c19145c20f7c4
Comment out /run/systemd confine for Fedora puppet

Upstream puppet has a constraint on matching systemd that /run/systemd
has to be around; this fails when building with dib in a chroot where
we're not actually running.

A full solution has to take into account Debian and multiple init
systems (see linked issues).  For the moment, just comment it out.

Change-Id: I6e4832caf6162a67408c34151b6e1d641a75fb8b