Updated openstack/openstack Project: openstack/diskimage-builder  a9c8e6d55d5371a22e502ebfa173b9d8e72ed0d1
Load the 8021q kernel module in simple-init

Vlan support was recently added to glean. However, if the 8021q module
is not loaded, glean will fail to bring up a tagged interfaced defined
in /etc/network/interfaces.d/. Manually attempting to bring up the
interface results in an error[1]. This patch ensures that the 8021q
module is loaded so that tagged interfaces can be brought up at boot.

[1] http://paste.openstack.org/show/480027/

Change-Id: I15d805c07d4b5e1161d831f0393d027e4325137f