Updated openstack/openstack Project: openstack/nova  eeb28b66ccbe980af3a9bd617246a90522295c53
tighten regex on objectify

There are some strange and wonderous work arounds in the json loader
for samples templates to do with templated values which are ints. This
was always kind of a hack, but when attempting to add a replacement
value in the middle of a string it was clear that the anchoring for
this regex was wrong.

Update the begin anchor to be ': ' instead of [^"], which would match
anywhere inside a string.

Change-Id: I9811b34d58a4d4cf7f181689b25212893880dcf4