Updated openstack/openstack Project: openstack/python-novaclient  d336385dd2972569b0e1b0c7ca780a32ce84ff27
Fix H404/405 violations in novaclient/tests/*

There is a lot of H404/405 violations in novaclient, and it is better
to fix those to have a better doc string for class/methods.

This patch fixes these violations for files under  novaclient/tests/
folder.

As there are lot of violations and cannot be fixed in single patches,
So separating those in multiple patches for easy review.

Partial-Bug: #1521899
Change-Id: I6e712ece14c745013bfba0bee9d77e7875dd2263