Updated openstack/openstack Project: openstack/keystone  f45af263f1240e9b0ec9c1695b586775a289f00d
Pass dict into update() rather than **kwargs

It looks like in the keystone/tests/core.py file 'ref', which is
a dictionary declared in most of the methods, is being updated by either
passing a dictionary or passing kwargs. To me it doesn't make
sense to make keyword arguments out of a dictionary, to wind up
adding them to another dictionary. This patch changes these so
update() is called passing in a dictionary instead of keyword args.

e.g. dict.update(dict1)

Change-Id: Iaec993c4728e82d6b27d8862005fa319daae3bb0