Updated openstack/openstack Project: openstack/requirements  6a1e28d92da5f604061f7d4849cae09e9eec4ade
Cap elasticsearch client <2.0

The elasticsearch 2.0 python client is not backwards compatible
with 1.x, nor with server versions prior to 2.0 (and similarly, the
1.x client cannot be used with server versions after 1.x). See
https://elasticsearch-py.readthedocs.org/en/master/ for details.

Since devstack installs elasticsearch 1.x it seems sensible to cap
the client version. Deployers using version 2.0 will need to install
that client version in projects requiring it.

Change-Id: I78c18185a6cdaf3acf51e84c5475aa6ff29c5e1f
Partial-Bug: #1515412