Updated openstack/openstack Project: openstack/oslo.config  1a12643dc65555a63444728c3db55a82c0a2ebe5
Test equality of types.String choices as a set

The equality of a String might return false if the choices value
is a List for one instance and Tuple for another.  Also the order
can prevent a valid match.  This patch normalizes the compare with
a set to ensure equality in those cases.

Change-Id: Ic5c36c390401c4cd60b1c5c46be93744b0025f06
Closes-Bug: #1523792