Updated openstack/openstack Project: openstack/oslo.utils  d423b3b74cfae284c4b46d3bad5232e3840e05f3
Expose function signature fetching function

This function is the recommended way to get information
about functions going forward, so all the variations of
getargspec can just use this instead (including the variations
of this used in openstack).

This exposes a 'get_signature' function that correctly aliases
the exposed function in python3.x and the backwards compat
library (funcsigs) that exposes this same information in python2.x

Internally in the other reflection functions we now switch
away from using getargspec as well, and replace it with using
this new signature object/function instead.

Depends-On: I1aa0054089ca57fc2d68779f4ee133a9750fec2a

Change-Id: I910b353f5db290832ef87bf7c5e5bb2e9788e3ec