Updated openstack/openstack Project: openstack/ossa  75267d110b96b0c2551dc635d7202320b6b2c726
Add class B3 and an example for C1 to the taxonomy

Be more explicit that OSSA does not cover vulnerabilities in
experimental features, backends and drivers by adding a new class B3
for these in the taxonomy. Also clarify that vulnerabilities relying
on UUID guessing are considered impractical, as an example for class
C1.

Change-Id: Ie73dfb0358913e6bdfeba56e6105f8156382d042