Updated openstack/openstack Project: openstack/requirements  1b61fb90f34dce91c60f90ad00d82b8ae1d7b418
Block oslo.messaging 3.1.0 since it breaks Ironic

oslo.messaging change I21d76abf1c30b9e89cb8a6c20d1c0cc79cd83a3f was
released in 3.1.0 and due to debtcollector bug 1520397 it breaks Ironic
loading it's request context serializer.

That was worked around in Ironic on master (mitaka) in commit
37d36f1536944de38a0f7a726e625b893088e1be but that doesn't fix Ironic in
stable/liberty, and we aren't going to backport that workaround to
stable/liberty.

So we need to fix oslo.messaging (either revert the deprecation change
or remove the usage of debtcollector) and release a fix there to fix
stable/liberty, but we still have to block the broken 3.1.0 release.

Related-Bug: #1524041

Change-Id: I3c5a9851ada30b6d7470113014b542b550a4ae8b