Updated openstack/openstack Project: openstack/neutron  143ef92e488cce39c1153af92ce0808babfcffb2
Revert "Revert "OVS agent reacts to events instead of polling""

The original change had to be reverted because it caused
tests failing in the gate. The failures were due to the
fact that when a port was not ready, an exception was
thrown to trigger a full resync of the agent. This
behavior was meant to be temporary and was fixed in a
dependent patch that was not merged though. This revert
moves the handling of not ready ports in this patch.
It also refactors the unit tests a bit.

This reverts commit e7270d9505fe09ec6d687e7796a814bf883b5416.

Change-Id: I6574cef3c95525ace6a98cf968ee159190681394