Updated openstack/openstack Project: openstack/neutron  a9b5398083556d4db6f29ee2734ec279c480306b
Make keepalived add_vip idempotent

L3 agent operations are designed to be idempotent to be resiliant
in the face of missed notifications, out of order notifications
processing and agent restarts. One exception is keepalived's
add_vip that was made to explicitly barf when the agent tries
to add a vip that was already inserted to keepalive'd internal
cache. This was done in patch 142630 by my request to root
out bugs. Well, we found such a bug and in retrospect may not
have been the smartest idea.

Change-Id: Iead975787e2847828286d7b644dcbe33cf57ace9
Related-Bug: #1523999