Updated openstack/openstack Project: openstack/heat  1ea579779ecb6d541561a9ba11ba62633a7094bc
Call _persist_state for UPDATE in 'state_set'

UPDATE flow doesn't call 'state_set', to set the state.
However, this is called in 'set_stack_and_resource_to_failed',
to set the stack status to FAILED. This includes reset during
stack UPDATE.

With https://review.openstack.org/#/c/245810/5, we removed UPDATE
from persisting the state as part of releasing the lock. This has
resulted in the state not being persisted for UPDATE when
'reset_stack_status' is called.

Change-Id: I8a80e329604fdec6498acca0a71723100d412384
Closes-Bug: #1523717