Updated openstack/openstack Project: openstack/neutron  8e9a737801b4829e5c2359d2b1445c36764535b5
devref: Rolling upgrade mechanism for rpc-callbacks

Versioned Object push notifications require the server to be aware
of supported versions in the agents, since they are subscribed
to neutron-vo-<resource-type>-<version-number>.

During upgrade time, the server would need to downgrade and serialize
the objects across version subset, and send it to the fanout
queues for agent consumption.

One manual solution could be manual admin pinning, but we can do
better than that, making administrator lives easier if we provide
a reliable mechanism for remote version auto discovery.

Change-Id: I02b694137eb2d58e5f2f3e7631f0e4b90f7c17ad