Updated openstack/openstack Project: openstack/heat  28d4d39372bfe16d67b2765b57a37b0a4f811573
Add validate_with_client in BaseCustomConstraint

Since we replace not found error with EntityNotFound instead,
we can now consider use a common function for validate_with_client in most
of constraints. For some constraints that have a more complex validation will
overwrite the function(This is what we already used).

Change-Id: I1c554ba7848279cd60db1661fe7ebbd5b72235d8