Updated openstack/openstack Project: openstack/keystone  44c1d2f953a966723af1e6b14f6727b9282a907b
Fix exposition of bug about limiting with ldap

The test did not take into account that even with the limits set,
as part of some internal calculation we want to get the full list
of entities. Calling the API without a hints list should achieve
this, and return all available entries.

Drop the custom test and use generic LimitTests.

Change-Id: Id3120ed66cc23228edcf18e0177c1df1955ac5e7
Partial-Bug: 1501698