Updated openstack/openstack Project: openstack-dev/devstack  9fc3ba408a97d0dd39ce26dd8dbcdb3b110cde71
Ironic: add flag for using plugin

This adds a flag to skip ironic code if the ironic devstack plugin is in
use. This flag will be set to true in ironic's devstack plugin to
indicate that the plugin should be in control, rather than devstack.

This is for the transition period only, and will be removed with the
rest of the ironic code in the devstack tree, once the gate
is configured to use the ironic plugin.

Change-Id: Id01d97fd13fa9f866d645ec5077834ddb78b2b89