Updated openstack/openstack Project: openstack/python-swiftclient  4d9b1e760d48aba570ccf07281ddaad110ed9a6a
Remove py26 support

As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.
Closes-Bug: 1519510
Depends-On: I37116731db11449d0c374a6a83a3a43789a19d5f

Change-Id: I776847ce77dfe82880f34d0b7804514e5aed3f8d