Updated openstack/openstack Project: openstack/nova  f660a8068f3f90231d7c7acb3f7c44c1d0e26c7d
Cache the automatic version pin to avoid repeated lookups

The compute rpcapi module was looking up the desired automatic
version pin each time it was initialized. Turns out, we create and
destroy this object too often, which means we do a lot of those
lookups unnecessarily. That's a larger problem that will be harder
to solve, but we can be smarter about the DB impact of that.

This adds a cache of the last version we discovered and avoids doing
the lookup when possible. It makes the reset() method invalidate the
cache prior to reinitializing the object so that we will do the lookup
on SIGHUP as desired.

Closes-Bug: #1524444
Change-Id: Ia1c58efa7105d32973f6921c2144b17c30b764c2