Updated openstack/openstack Project: openstack/oslo.messaging  0a0e6d0d50ac2a1f937ba45da2d54bec01b8c4ab
notif: Check the driver features in dispatcher

The transport/driver features check is done into the get listener
methods.
So when these methods are not used the driver features checks is not
done.

This change moves it into the dispatcher layer to ensure the
requirements are always checked.

This changes a bit the behavior of when the check occurs. Before
it was during the listener object initialisation. Now this
when the listener server start.

Change-Id: I4d81a4e8496f04d62e48317829d5dd8b942d501c