Updated openstack/openstack Project: openstack/tempest  02ef4c2956318b09cc47506af2361ab3d8907a10
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34
we have to run first the py34 env to, then, run py27. This patch
puts py34 first in the tox.ini list of envs to avoid this problem
to happen.
Closes-bug: #1489059
Change-Id: Ia11ecf7bd147e35b7c8ad3db0eaad17f893e78ba