Updated openstack/openstack Project: openstack/nova  2e46e453d8fe377f3823df1572d9fc08f03f0964
Report compute-api bugs against nova

For compute-api set the launchpad project users report bugs
against to "nova". Users can report bugs using the "bug icon" that will
directly link to the launchpad project, it currently goes to
"openstack-manuals" which is wrong for this content.

This variable is used by openstackdocstheme 1.2.6.

Closes-Bug: #1524476
Change-Id: I08ba11bb1c5388bc4611cea581274ac5d724c8d9