Updated openstack/openstack Project: openstack-infra/shade  5e83629fc1aa4d814a498b484b9c60d755e0eb2f
Make a new swift client prior to each image upload

Since swift clients do not support keystone sessions any client that
exists for a period of time runs the risk of having an expired token and
the OpenStackCloud object will never recreate it to reauth. Work around
this by making a new client whenever we attempt image uploads to swift.

There are other potential ways to address this. We could pass the auth
info into swift_client.get_capabilities() so that swift will reauth on
its own if it needs to. We could just always make a new swiftclient for
every swiftclient operation. We could add keystone session support to
swiftclient. This workaround is simple and uses existing APIs with shade
and os-c-c so starting with it.

Change-Id: Ib35dd49627bc2209060848e719e3cec40dbb4f2a