Updated openstack/openstack Project: openstack/tempest  2d4f47fac7ed1ae60ee41f86cb2c5535387c16c8
Change docstring of compute client

We need to put the links to api-site for explaining what parameters
can be passed to Nova, then this patch adds them.
In addition, this patch changes docstring for the consistency.

NOTE: This patch changes the methods which call "/servers" only.
      The other methods like "/servers/<id>/action/" will be changed
      with the other patches because servers_client is big and it is
      difficult to change/review all of them in a single patch.

Partially implements blueprint consistent-service-method-names

Change-Id: I36b94d6b9f136ca65cb361a8af65966f21b86b03