Updated openstack/openstack Project: openstack-infra/puppet-gerrit  6d6244e831f197e9efebe56e8059b666a06b263d
Allow puppet to configure the Gerrit drafts feature

In Gerrit 2.11 disabling draft pushes in gerrit ACLS was not enough to
disable the drafts feature.  There is now a change.allowDrafts [1]
settting in gerrit.config to disable drafts which also needs to be
set.  This change allows puppet configure this setting.

[1] https://review-dev.openstack.org/Documentation/config-gerrit.html#change

Change-Id: Iea970a85d433d8707102a3dcaf8cbae2d5662b6e