Updated openstack/openstack Project: openstack/nova  20db45da9945e4c8125619df1fee69cf6663f1c0
default host to service name instead of uuid

The host associated with a service looks like it really should be a
hostname and not a uuid in the current code. This changes it to
default to the name of the service started if nothing else is
specified in the test fixture. It then updates all samples to expect a
thing that looks like a name, and not a uuid.

This also allows us to drop the special casing of this function for
aggregate testing.

Change-Id: I37415b89b733c021daa38040fee447cf3934b821