Updated openstack/openstack Project: openstack-infra/shade  8d5abfbf56a5620d1326d44dee8ffdc9dfba1c62
Bug fix: delete_object() returns True/False

Our delete APIs return True if the delete succeeded, or False if
the thing being deleted was not found. delete_object() was not doing
this, so this makes it consistent with the other delete API calls.

Also adds missing unit tests for this method.

Change-Id: I0951765193459300f08b0ab804e6ca327c6fa57d