Updated openstack/openstack Project: openstack-infra/shade  fb8ea73f27f683519dcb27ed9e83cd6e4e423379
Stack API improvements

The exception that would be raised from list_stacks() if it failed
would not have any type of error message (except the inner exception).
This could be confusing, so add some text that indicates where the
exception is being thrown. Unit tests for this method are added.

Also improve some comments in delete_stack() and add unit tests
for this method.

Change-Id: I979a2948938fc73b708c26d81599061bac7681d4