Updated openstack/openstack Project: openstack/neutron  dbc541be5428e092315ed3b51fa2295a5ea49068
Add address scope to ports in RPC response to L3 agent

The L3 agent needs to know the address scope of each port of each
router it sets up in order to enforce isolation between scopes.

This commit adds a devref for the address scopes and subnet pools
features.

Change-Id: I6a7b3708fadefff1919d70ab1b8bc345b3fbe81c
Partially-Implements: blueprint address-scopes