Updated openstack/openstack Project: openstack/horizon  9eae7d16bfaedc06be629bcc55d17aa8866b0cb5
Add extensions to $q for resolving all promises

This patch provides an extension service for the $q
service which waits for all promises to settle and then
return which promises fail and which promises pass.

This was a result of having similar looking code in
https://review.openstack.org/#/c/231335/
and
https://review.openstack.org/#/c/234408/

The method name was inspired method which exists in $q
from Kris Oswal but not in the Angular implementation of it.
https://github.com/kriskowal/q/blob/v1/q.js#L1676

Co-Authored-By: Errol Pais<epais@thoughtworks.com>
Co-Authored-By: Kyle Olivo<keolivo@thoughtworks.com>

Change-Id: Ic3d36bba1886e06d7e7aee50c1ddfa17b51a1a65
Partially-Implements: blueprint angularize-images-table