Updated openstack/openstack Project: openstack-infra/project-config  f83c2866ce9edb783e62b57d373681527e47c8f9
Enable devstack-plugin-ceph NV jobs for glance and nova

Commit a53859db32475c948ee68ce1f838b20e773b8687 enabled the
devstack-plugin-ceph NV job for cinder only as we didn't
want to load the infra during the Mitaka-1 release rush. This
patch now adds the job to glance and nova projects.

This patch is part of a larger effort of moving the devstack-ceph
CI job from devstack hook -> plugin method.

Change-Id: I6acd66a0a4c43e2af3889928a9751ae3b4993793