Updated openstack/openstack Project: openstack-infra/subunit2sql  16087b0e19af90e8f95d8a1f136ef3589158b34d
Enable db api tests on all backends

Previously the db api tests were added using testscenarios with the
plan to expand them to run on all the supported backends. This commit
implements just that by adding support for using all 3 supported
backends. As part of this change it also makes the tests properly
opportunistic so they skip rather than fail if a properly configured
db backend is not available.

Change-Id: I0a772b38f4248425f44edac96ad3de98391f6ae3