Updated openstack/openstack Project: openstack/api-site  1fdef1f26c5e7c988ff105869111065ad45383a1
Mark LBaaS v2 API as stable and LBaaS v1 API as deprecated

Neutron Liberty release note says:
The LBaaS V2 API is no longer experimental. It is now stable.
The LBaaS V1 API is marked as deprecated and is planned to be removed
in a future release. Going forward, the LBaaS V2 API should be used.

Change-Id: I0cc2349acbbf02845670f2e13f1d56659722c8d9
Closes-Bug: #1525567