Updated openstack/openstack Project: openstack/cinder  a91a85599b855626e89771e6f626edca10e44681
Storwize: add config option to control flash copy rate

When creating a flash copy of a vdisk during a clone
or a snapshot operation the svcinfo mkfcmap operation
allows the specification of a copy rate for the flash copy.
The default copy rate is 50. The issue with 50 is it not
always an acceptable default copy rate for a particular
operator. Some operators want it slower but some want to
increase the default copy rate.

Adding a configuration option storwize_svc_flashcopy_rate
to allow an operator to specify the default flash copy
rate to be used when starting a vdisk flash copy on a
clone or snapshot operation.

DocImpact

Change-Id: Iffa50cf934f5d340a1f24ed3cfea8381b6dd6630
Closes-Bug: 1517185