Updated openstack/openstack Project: openstack/cinder  d677acba39c934072439a806869c04fe69f65acc
NetApp: E-Series fix JSONDecodeError on first add

The NetApp E-Series driver would throw a JSONDecodeError on the
addition of a new storage-system. This would occur because a
JSON response was expected, but text/html was being returned as
the Content-Type when a storage-system has been added but is still
coming online.

This patch adds a retry to the driver startup to wait for an array
to come online. It also adds code to avoid the JSONDecodeError
and fail based on the HTTP Response's return code.

Change-Id: I405723039c7a0df10d39bd2eb961ee64d71c2d01