Updated openstack/openstack Project: openstack/nova  aaf6ab4b78a8412bc56cfbe6a4387e634ad03e89
Do not update links for 'versions' tests

Now we have logic in api sample base class is to update the
api version in links for sample file data. So that single set
of sample files works for all tests scenario(v2, v2.1, v2 comp and
microversion etc).

There are some cases where we do not need to update the links.
Example-
'Version' API tests does not run on particular endpoint and
there we need to skip the links update logic.

This patch provides options for tests to update the links or not.

Change-Id: Ic446e297b484b967098cae61f50700c552d98134