Updated openstack/openstack Project: openstack/cinder  892de6d0a91c6750ed9869a25af8924b0fcde0a6
Recognize extra fields in CinderObjectDictCompat

We're still in the process of transitioning to the object notation in a
lot of places and that's why we've decided to keep specialized get in
CinderObjectDictCompat. This one wasn't looking at aliases (properties)
from obj_extra_fields. This commit fixes that and adds a unit test for
this method.

We should eventually stop using it, but for Mitaka this will make a lot
of things easier.

Change-Id: I1df610bedffdeae56444ce0a40f8fb48216e91b7
Related-Bug: 1516903